Patient Detials ui updated
