+++ lib/departments/second.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'dart:convert';
 import 'package:http/http.dart' as http;
 
-import 'package:flutter_app_json_local/doctor_details.dart';
+import 'package:flutter_app_json_local/doctorDetails.dart';
 import 'package:flutter_app_json_local/customWidgets/drawer.dart';
 import 'package:flutter_app_json_local/customWidgets/doctorsList.dart';
